refactor(run.Store): move run event management from run.Store to Orchestrator
Applying a lesson learned while working on the new collection subsystem: if we
can avoid interfaces requiring an event bus or publisher, it's worth doing so.
Move event handling logic up into the orchestrator where it only needs to be
written once, and drop the requirements that run.Store implementations take a
bus. This drops three methods from the run.Store interface

Add an extension interface for optimized event writing, and check for that
interface within orchestrator run ID event subscriptions. Move event persistence
spec tests up into an orchestrator test.
