+++ automation/orchestrator_test.go
@@ -39,7 +39,7 @@ func TestIntegration(t *testing.T) {
 
 	bus := event.NewBus(ctx)
 
-	runStore := run.NewMemStore(bus)
+	runStore := run.NewMemStore()
 	workflowStore := workflow.NewMemStore()
 	opts := OrchestratorOptions{
 		WorkflowStore: workflowStore,
@@ -105,4 +105,173 @@ func errOnTimeout(t *testing.T, c chan string, errMsg string) <-chan struct{} {
 		done <- struct{}{}
 	}()
 	return done
+}
+
+func TestRunStoreEvents(t *testing.T) {
+	// mock time
+	prevNow := event.NowFunc
+	defer func() { event.NowFunc = prevNow }()
+
+	timestamps := []time.Time{}
+	totalEventsEmitted := 8
+	eventNumber := 0
+	for i := 0; i < totalEventsEmitted; i++ {
+		t := time.Unix(int64(i), 0)
+		timestamps = append(timestamps, t)
+	}
+	event.NowFunc = func() time.Time {
+		t := timestamps[eventNumber]
+		eventNumber++
+		return t
+	}
+
+	timestampNum := 0
+	nextTimestamp := func() *time.Time {
+		if timestampNum >= len(timestamps) {
+			t.Fatal("timestamp error, out of bounds")
+		}
+		t := timestamps[timestampNum]
+		timestampNum++
+		return &t
+	}
+
+	ctx := context.Background()
+	bus := event.NewBus(ctx)
+	runStore := run.NewMemStore()
+	workflowStore := workflow.NewMemStore()
+	wf, err := workflowStore.Put(&workflow.Workflow{
+		DatasetID: "dataset_id",
+		OwnerID:   "owner_id",
+		Created:   &time.Time{},
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+	r := &run.State{WorkflowID: wf.ID}
+
+	// this runFunc simulates events emitted by the transform package
+	runFuncFactory := func(ctx context.Context) Run {
+		return func(ctx context.Context, streams ioes.IOStreams, w *workflow.Workflow, runID string) error {
+			r.ID = runID
+
+			// event 0
+			bus.PublishID(ctx, event.ETTransformStart, r.ID, nil)
+			ts := nextTimestamp()
+			r.Status = run.RSRunning
+			r.StartTime = ts
+			confirmStoredRun(t, runStore, r)
+
+			// event 1
+			bus.PublishID(ctx, event.ETTransformStepStart, r.ID, event.TransformStepLifecycle{
+				Name:     "step start",
+				Category: "step start category",
+			})
+			ts = nextTimestamp()
+			expectedStep := &run.StepState{
+				Name:      "step start",
+				Category:  "step start category",
+				Status:    run.RSRunning,
+				StartTime: ts,
+			}
+			r.Steps = append(r.Steps, expectedStep)
+			confirmStoredRun(t, runStore, r)
+
+			// event 2
+			bus.PublishID(ctx, event.ETTransformPrint, r.ID, "transform print")
+			ts = nextTimestamp()
+			expectedPrintEvent := event.Event{
+				Type:      event.ETTransformPrint,
+				Timestamp: (*ts).UnixNano(),
+				SessionID: r.ID,
+				Payload:   "transform print",
+			}
+			r.Steps[0].Output = []event.Event{expectedPrintEvent}
+			confirmStoredRun(t, runStore, r)
+
+			// event 3
+			bus.PublishID(ctx, event.ETTransformError, r.ID, "transform error")
+			ts = nextTimestamp()
+			expectedErrorEvent := event.Event{
+				Type:      event.ETTransformError,
+				Timestamp: (*ts).UnixNano(),
+				SessionID: r.ID,
+				Payload:   "transform error",
+			}
+			r.Steps[0].Output = []event.Event{expectedPrintEvent, expectedErrorEvent}
+			confirmStoredRun(t, runStore, r)
+
+			// event 4
+			bus.PublishID(ctx, event.ETTransformDatasetPreview, r.ID, "transform dataset preview")
+			ts = nextTimestamp()
+			expectedDatasetPreviewEvent := event.Event{
+				Type:      event.ETTransformDatasetPreview,
+				Timestamp: (*ts).UnixNano(),
+				SessionID: r.ID,
+				Payload:   "transform dataset preview",
+			}
+			r.Steps[0].Output = []event.Event{expectedPrintEvent, expectedErrorEvent, expectedDatasetPreviewEvent}
+			confirmStoredRun(t, runStore, r)
+
+			// event 5
+			bus.PublishID(ctx, event.ETTransformStepStop, r.ID, event.TransformStepLifecycle{Status: "succeeded"})
+			ts = nextTimestamp()
+			expectedStep = r.Steps[0]
+			expectedStep.StopTime = ts
+			expectedStep.Status = run.RSSucceeded
+			expectedStep.Duration = int(expectedStep.StopTime.Sub(*expectedStep.StartTime))
+			confirmStoredRun(t, runStore, r)
+
+			// event 6
+			bus.PublishID(ctx, event.ETTransformStepSkip, r.ID, event.TransformStepLifecycle{Name: "step skip", Category: "step skip category", Status: "skipped"})
+			ts = nextTimestamp()
+			expectedSkipStep := &run.StepState{
+				Name:     "step skip",
+				Category: "step skip category",
+				Status:   run.RSSkipped,
+			}
+			r.Steps = append(r.Steps, expectedSkipStep)
+			confirmStoredRun(t, runStore, r)
+
+			// event 7
+			bus.PublishID(ctx, event.ETTransformStop, r.ID, event.TransformLifecycle{Status: "succeeded"})
+			ts = nextTimestamp()
+			r.StopTime = ts
+			r.Status = run.RSSucceeded
+			r.Duration = int(r.StopTime.Sub(*r.StartTime))
+			confirmStoredRun(t, runStore, r)
+
+			return nil
+		}
+	}
+	applyFuncFactory := func(ctx context.Context) Apply {
+		return func(ctx context.Context, streams ioes.IOStreams, w *workflow.Workflow) error {
+			return nil
+		}
+	}
+
+	opts := OrchestratorOptions{
+		WorkflowStore: workflowStore,
+		RunStore:      runStore,
+	}
+	o, err := NewOrchestrator(ctx, bus, runFuncFactory, applyFuncFactory, opts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer o.Shutdown()
+
+	if err := o.RunWorkflow(ctx, wf.ID); err != nil {
+		t.Fatal(err)
+	}
+}
+
+func confirmStoredRun(t *testing.T, s run.Store, expect *run.State) {
+	t.Helper()
+	got, err := s.Get(expect.ID)
+	if err != nil {
+		t.Fatalf("getting run: %s", err)
+	}
+
+	if diff := cmp.Diff(expect, got); diff != "" {
+		t.Errorf("stored run mismatch: (-want +got):\n%s", diff)
+	}
 }
