+++ automation/run/store.go
@@ -1,7 +1,6 @@
 package run
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"sort"
@@ -12,8 +11,6 @@ import (
 )
 
 var (
-	// ErrNilRun indicates that the run.State is nil
-	ErrNilRun = fmt.Errorf("nil run")
 	// ErrNotFound indicates that the run.State was not found in the Store
 	ErrNotFound = fmt.Errorf("run not found")
 	// ErrUnknownWorkflowID indicates that the given workflow.ID has no
@@ -47,15 +44,18 @@ type Store interface {
 	// ListByStatus returns a list of run.State entries with a given status
 	// looking only at the most recent run of each Workflow
 	ListByStatus(s Status, limit, offset int) ([]*State, error)
+}
 
-	// SubscribeID subscribes to events emitted w/ the given run id. Uses
-	// `AddTransformEvent` to interpret/change the runState, and calls
-	// Update on the Store
-	SubscribeID(id string) error
-	// Unsubscribe ID unsubscribes to events emitted w/ the given run id
-	UnsubscribeID(id string) error
-	// Bus returns the bus that the Store subscribes to
-	Bus() event.Bus
+// EventAdder is an extension interface that optimizes writing an event to a run
+// Result should equal to calling:
+//   run := store.Get(id)
+//   run.AddTransformEvent(e)
+//   store.Put(run)
+type EventAdder interface {
+	Store
+	// AddEvent writes an event to the store, attaching it to an existing stored
+	// run state
+	AddEvent(id string, e event.Event) error
 }
 
 // MemStore is an in memory representation of a Store
@@ -82,33 +82,27 @@ func newWorkflowMeta() *workflowMeta {
 var _ Store = (*MemStore)(nil)
 
 // NewMemStore returns a MemStore
-func NewMemStore(bus event.Bus) *MemStore {
+func NewMemStore() *MemStore {
 	s := &MemStore{
-		bus:       bus,
 		workflows: map[workflow.ID]*workflowMeta{},
 		runs:      map[string]*State{},
 	}
-	s.handleRunEvents = HandleRunEventsFactory(s)
 	return s
-
 }
 
 // Put adds a run.State to a MemStore
 func (s *MemStore) Put(r *State) (*State, error) {
 	if r == nil {
-		return nil, ErrNilRun
+		return nil, fmt.Errorf("run is nil")
 	}
 	run := r.Copy()
 	if run.ID != "" {
-		fetchedR, err := s.Get(run.ID)
-		if errors.Is(err, ErrNotFound) {
-			return nil, ErrNotFound
+		if fetchedR, err := s.Get(run.ID); !errors.Is(err, ErrNotFound) {
+			if fetchedR.WorkflowID != run.WorkflowID {
+				return nil, ErrPutWorkflowIDMismatch
+			}
 		}
-		if fetchedR.WorkflowID != run.WorkflowID {
-			return nil, ErrPutWorkflowIDMismatch
-		}
-	}
-	if run.ID == "" {
+	} else {
 		run.ID = NewID()
 	}
 	if err := run.Validate(); err != nil {
@@ -265,41 +259,24 @@ func (s *MemStore) ListByStatus(status Status, limit, offset int) ([]*State, err
 	return set.Slice(start, end), nil
 }
 
-// Bus returns the event.Bus
-func (s *MemStore) Bus() event.Bus {
-	return s.bus
-}
+// AddEvent writes an event to the store, attaching it to an existing stored
+// run state
+func (s *MemStore) AddEvent(id string, e event.Event) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
 
-// SubscribeID subscribes to events emitted w/ the given run id. Uses
-// `AddTransformEvent` to interpret/change the runState, and calls
-// Update on the Store
-func (s *MemStore) SubscribeID(id string) error {
-	s.bus.SubscribeID(s.handleRunEvents, id)
-	return nil
-}
-
-// UnsubscribeID unsubscribes the event bus from the run id
-// TODO(ramfox): UnsubscribeID requires a bus.UnsubscribeID method
-func (s *MemStore) UnsubscribeID(id string) error {
-	return fmt.Errorf("UnsubscribeID is not yet implemented")
-}
-
-// HandleRunEventsFactory returns a handler that will properly handle run
-// events for the given store
-func HandleRunEventsFactory(s Store) event.Handler {
-	return func(ctx context.Context, e event.Event) error {
-		id := e.SessionID
-		run, err := s.Get(id)
-		if err != nil {
-			return fmt.Errorf("store.SubscribeID - id %q: error fetching run.State, %w", id, err)
-		}
-		if err := run.AddTransformEvent(e); err != nil {
-			return fmt.Errorf("store.SubscribeID - id %q: error adding transform event to run, %w", id, err)
-		}
-		_, err = s.Put(run)
-		if err != nil {
-			return fmt.Errorf("store.SubscribeID - id %q: error updating run.State, %w", id, err)
-		}
-		return nil
+	run, ok := s.runs[id]
+	if !ok {
+		return ErrNotFound
 	}
+	if err := run.AddTransformEvent(e); err != nil {
+		return fmt.Errorf("adding transform event to run: %w", err)
+	}
+
+	if _, ok = s.workflows[run.WorkflowID]; !ok {
+		wf := newWorkflowMeta()
+		s.workflows[run.WorkflowID] = wf
+	}
+	s.runs[id] = run
+	return nil
 }
